backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
gitpython==3.1.43
google-generativeai==0.7.2
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
pydantic==2.7.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
firebase-admin==6.5.0
